# docs/APIDocumentation.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ZooD Industries",
    "description": "Свободная соцсеть.",
    "version": "0.8.0"
  },
  "servers": [
    {
      "url": "http://localhost:888",
      "description": "Development"
    },
    {
      "url": "https://api.zood.fun",
      "description": "Production"
    }
  ],
  "paths": {
...
    "/api/profiles": {
      "get": {
        "tags": [
          "Profiles"
        ],
        "summary": "Получить список своих профилей",
        "operationId": "get_api_profiles_list",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Успешно",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProfileDTO"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "Profiles"
        ],
        "summary": "Изменить данные профиля",
        "operationId": "put_api_profiles_update",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdateCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешно"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Profiles"
        ],
        "summary": "Создать новый профиль",
        "operationId": "post_api_profiles_create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProfileCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешно"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ProfileDTO": {
        "properties": {
          "id": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string",
            "nullable": true
          }
        },
        "type": "object"
      },
      "ProfileUpdateCommand": {
        "properties": {
          "profileId": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string",
            "nullable": true
          }
        },
        "type": "object"
      },
      "CreateProfileCommand": {
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string",
            "nullable": true
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "bearerFormat": "JWT",
        "scheme": "bearer"
      }
    }
  },
...
}